+++ src/Core/Services/IStripeSyncService.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace Bit.Core.Services
+{
+    public interface IStripeSyncService
+    {
+        Task<bool> UpdateCustomerEmailAddress(string gatewayCustomerId, string emailAddress);
+    }
+}
