+++ src/Core/Services/Implementations/StripePaymentService.cs
@@ -9,7 +9,6 @@ using Bit.Core.Exceptions;
 using Bit.Core.Models.Business;
 using Bit.Core.Repositories;
 using Microsoft.Extensions.Logging;
-using Stripe;
 using StaticStore = Bit.Core.Models.StaticStore;
 using TaxRate = Bit.Core.Entities.TaxRate;
 
@@ -1429,23 +1428,7 @@ namespace Bit.Core.Services
 
             return createdCustomer;
         }
-
-        public async Task<bool> UpdateCustomerEmailAddress(string gatewayCustomerId, string emailAddress)
-        {
-            try
-            {
-                var customer = await _stripeAdapter.CustomerGetAsync(gatewayCustomerId);
-
-                await _stripeAdapter.CustomerUpdateAsync(customer.Id,
-                    new Stripe.CustomerUpdateOptions { Email = emailAddress });
-                return true;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-
+        
         public async Task<bool> CreditAccountAsync(ISubscriber subscriber, decimal creditAmount)
         {
             Stripe.Customer customer = null;
