+++ src/Core/Services/Implementations/UserService.cs
@@ -52,6 +52,7 @@ namespace Bit.Core.Services
         private readonly IOrganizationService _organizationService;
         private readonly IProviderUserRepository _providerUserRepository;
         private readonly IDeviceRepository _deviceRepository;
+        private readonly IStripeSyncService _stripeSyncService;
 
         public UserService(
             IUserRepository userRepository,
@@ -81,7 +82,8 @@ namespace Bit.Core.Services
             IGlobalSettings globalSettings,
             IOrganizationService organizationService,
             IProviderUserRepository providerUserRepository,
-            IDeviceRepository deviceRepository)
+            IDeviceRepository deviceRepository,
+            IStripeSyncService stripeSyncService)
             : base(
                   store,
                   optionsAccessor,
@@ -117,6 +119,7 @@ namespace Bit.Core.Services
             _organizationService = organizationService;
             _providerUserRepository = providerUserRepository;
             _deviceRepository = deviceRepository;
+            _stripeSyncService = stripeSyncService;
         }
 
         public Guid? GetProperUserId(ClaimsPrincipal principal)
@@ -566,7 +569,7 @@ namespace Bit.Core.Services
             if (user.Gateway == GatewayType.Stripe
                 && !string.IsNullOrWhiteSpace(user.GatewayCustomerId))
             {
-                var status = await _paymentService.UpdateCustomerEmailAddress(user.GatewayCustomerId,
+                var status = await _stripeSyncService.UpdateCustomerEmailAddress(user.GatewayCustomerId,
                     user.BillingEmailAddress());
 
                 if (!status)
